requirements.txt
xlrd
pandas
numpy
matplotlib
plotly
seaborn
statsmodels
scikit-learn
imblearn
xgboost
python-dotenv
pymongo[srv]
certifi
dill
pyYAML
scipy
kneed